script: Add script to help a user to manually create config
Add a script 'genusrcfg' to print the configuration content for a given
user. Users can use this script to get the config entry for a given
image.deb file for a given user config.

These can be copied and pasted in /boot/grub/custom/<user>.cfg.
User can then modify the kernel cmd line arguments, or change the
image/initrd.path or change name of the grub menu entry as per
requirement.

Signed-off-by: Ankit Nautiyal <akn_github@outlook.com>